Completion of entire Testbuilder exercise
This work takes an inputted string of a credit card number and returns
the name of the credit card network associated with the card (eg Visa,
Diner's Club, American Express, Switch, Mastercard, Discover, Maestro,
or China UnionPay).

Sorry - I completely forgot to commit my work incrementally on this
assignment.  I'll be sure to do so moving forward.  While this code
seems to be working and passing all tests, I believe there is some
room for refactoring improvement.  I'll go back and do so if there
is time.